requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
email-validator==1.1.1
Flask==1.1.2
Flask-MySQL==1.5.1
Flask-MySQLdb==0.2.0
flask-ngrok==0.0.25
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysqlclient==1.4.6
PyMySQL==0.9.3
python-dotenv==0.13.0
requests==2.24.0
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
